+++ yelp_kafka/error.py
@@ -45,3 +45,20 @@ class UnknownTopic(YelpKafkaError):
 
 class UnknownPartitions(YelpKafkaError):
     pass
+
+
+class OffsetCommitError(YelpKafkaError):
+
+    def __init__(self, topic, partition, error):
+        self.topic = topic
+        self.partition = partition
+        self.error = error
+
+    def __eq__(self, other):
+        if all([
+            self.topic == other.topic,
+            self.partition == other.partition,
+            self.error == other.error,
+        ]):
+            return True
+        return False