+++ CHANGELOG.md
@@ -4,3 +4,19 @@ This changelog only guarantees backward incompatible changes will be listed.
 ## backward incompatible changes
 - `yelp_kafka.monitoring.*` function signatures have been changed to
   (client, group, topics) and (client, group, topic, partitions).
+
+# v3.0.0 (2015-05-28)
+## backward incompatible changes
+- `yelp_kafka.monitoring.get_current_consumer_offsets`:
+  - function signature changed from (client, group, topics, fail_on_error)
+    to (client, group, topics, raise_on_error)
+  - moved to `yelp_kafka.offsets.get_current_consumer_offsets`
+- `yelp_kafka.monitoring.get_topics_watermarks`:
+  - function signature changed from (client, topics, fail_on_error)
+    to (client, topics, raise_on_error)
+  - moved to `yelp_kafka.offsets.get_topics_watermarks`
+- `yelp_kafka.monitoring.PartitionOffsets` namedtuple moved to
+  `yelp_kafka.offsets.PartitionOffsets`
+## added
+- `yelp_kafka.offsets` new functions added:
+  advance_consumer_offsets, rewind_consumer_offsets, set_consumer_offsets